listen after new function class completes
