+++ src/projects/climb/src/html/lobby/host.html
@@ -11,7 +11,7 @@
     <div>Player List</div>
     <div class="player-list"></div>
   </div>
-  <div class="start-game-wrapper hidden">
+  <div class="start-game-wrapper">
     <div>Press</div>
     <div class="button-wrapper">
       <button data-action="start_game">Start Game</button>
