+++ src/projects/climb/src/js/domain/class/Host/index.js
@@ -1,10 +1,13 @@
 import { createRoomCode } from './create-room-code';
-import { render } from '../../render';
 import { displayScreen } from '../../../library/display-screen';
+import { onPlayerListUpdated } from './on-player-list-updated';
+import { render } from '../../render';
+import { app } from '../../app';
 
 function Host() {
   const code = createRoomCode();
-  const ref = window.firebase.database().ref(`rooms/${code}`);  
+  const ref = window.firebase.database().ref(`rooms/${code}`);
+  let playerList = '';
   
   // remove room from firebase when disconnected
   ref.onDisconnect().remove();
@@ -22,13 +25,17 @@ function Host() {
 
   return {
     code,
+    listen,
+    playerList,
     ref,
   }
 }
 
-// function onPlayerListUpdated(e) {
-//   console.log('on player list updated', e);
-// }
+function listen() {
+  console.log('listening');
+  const { ref } = app.host;
+  ref.child('players').on('value', onPlayerListUpdated);
+}
 
 export {
   Host,
