+++ src/projects/climb/src/js/domain/class/Host/on-player-list-updated.js
@@ -1,0 +1,16 @@
+import { app } from '../../app';
+import { render } from '../../render';
+
+function onPlayerListUpdated(snapshot) {
+  console.log('on player list updated', snapshot);
+  app.host.playerList = snapshot.exists() ? snapshot.toJSON() : {};
+  
+  render({
+    playerType: 'host',
+    screen: 'lobby',
+  });
+}
+
+export {
+  onPlayerListUpdated,
+}
