--- src/projects/climb/src/js/domain/enter-room/game.js
@@ -1,8 +1,0 @@
-import app from "../app";
-// import enterRoom from '../enter-room';
-
-function Game() {
-  console.log('this is the game screen');
-}
-
-export default Game;
