--- src/projects/climb/src/js/domain/enter-room/index.js
@@ -1,17 +1,0 @@
-import lobby from './lobby.js';
-import game from './game.js';
-import displayScreen from '../../library/display-screen';
-
-function enterRoom(screen) {
-  const joinRoom = {
-    lobby,
-    game,
-  }
-
-  joinRoom[screen]();
-  displayScreen(screen);
-}
-
-export {
-  enterRoom,
-}
